Happy new Year!!

+++ common/functions/src/aggregates/adaptors/aggregate_null_unary_adaptor.rs
@@ -50,7 +50,7 @@ impl<const NULLABLE_RESULT: bool> AggregateNullUnaryAdaptor<NULLABLE_RESULT> {
     #[inline]
     pub fn set_flag(place: StateAddr) {
         if NULLABLE_RESULT {
-            let c = place.get::<usize>();
+            let c = place.get::<u8>();
             *c = 1;
         }
     }
@@ -58,15 +58,15 @@ impl<const NULLABLE_RESULT: bool> AggregateNullUnaryAdaptor<NULLABLE_RESULT> {
     #[inline]
     pub fn init_flag(place: StateAddr) {
         if NULLABLE_RESULT {
-            let c = place.get::<usize>();
+            let c = place.get::<u8>();
             *c = 0;
         }
     }
 
     #[inline]
-    pub fn get_flag(place: StateAddr) -> usize {
+    pub fn get_flag(place: StateAddr) -> u8 {
         if NULLABLE_RESULT {
-            let c = place.get::<usize>();
+            let c = place.get::<u8>();
             *c
         } else {
             1
